Upgrade Maven + Eclipse to compile against 4.19 (2021-03) API
There is a new 'e419' profile which pulls bundles from the 2021-03
update site, because with the new AspectJ JDT files would no longer
compile with 'e410' bundles.

Because there are dependencies containing Java 11 class files, we now
have to build using JDK 11+, but target level 1.8 is still OK.

This also has the consequence of switching to Nick Wong's AspectJ Maven
fork, because the Mojohaus version is without a maintainer and only
supports Java 8.

Another important change is that I adjusted two manifests in order to
require more bundles, because otherwise an installed AJDT will not work
correctly in Eclipse 2021-03, even after the Java sources have been
adjusted.

Please note that the sources will not compile for this commit, because I
want to separate the source code adjustments to a subsequent commit. It
is easier to find and improve them later that way.

Signed-off-by: Alexander Kriegisch <Alexander@Kriegisch.name>