Add GitHub Maven build workflow + Dependebot
By default, a second build also running tests will be skipped, unless
the user interactively starts a workflow from the GitHub website and
manually selects a value of 'true' for the interactive 'Run tests'
option.

The main reason for skipping test execution by default is that tests are
slow, because most of them are UI tests, and that currently many tests
are failing, because they have not been maintained for a long time.

Signed-off-by: Alexander Kriegisch <Alexander@Kriegisch.name>