Refactor AspectJProjectWizardTest, solving problem with JMS module name
AspectJProjectWizardTest.testProjectWizardPerformFinish often made the
build time out, because the test was not unchecking the "Create
module-info.java" option in the new Java project wizard, which led to
flaky behaviour: a pop-up dialogue asking for a JMS module name often
popped up unexpectedly. So now the option is unchecked, solving the
problem.

While iteratively trying to find a solution, I also refactored both
AspectJProjectWizardTest and AspectJProjectWizardExtension.

Signed-off-by: Alexander Kriegisch <Alexander@Kriegisch.name>