Remove superfluous method AspectJProjectWizardExtension.useDefaultLocation
The method sets an option which in current Eclipse versions is the
default anyway. Hence, there is no need for more ugly reflection code.

I did not remove the method right away in the previous commit, because
after refactoring I wanted the method to be committed once, so it can be
restored easily in the future if necessary.

Signed-off-by: Alexander Kriegisch <Alexander@Kriegisch.name>