Remove AspectJ source and class files from AJDT repository
This is the follow-up for the previous commit. I split the commit in two
parts in order to facilitate code review. In the previous commit, there
are the actual changes to the build and .gitignore files. In this one,
I am simply deleting all the upstream files which are no longer needed
in the SCM and which were misplaced there from the beginning.

Signed-off-by: Alexander Kriegisch <Alexander@Kriegisch.name>