Remove VJET from .gitignore
VJET used to be an Eclipse plugin for Javascript developers, started at
https://github.com/ebayopensource/vjet, then migrated to
https://projects.eclipse.org/projects/webtools.vjet, where the last
release was in 2012, before it was retired. Now there is JSDT, see
https://www.eclipse.org/webtools/jsdt/. While also not being a very
active project, at least there is some activity, and it has not been
retired yet. So I would assume that whoever uses JS nowadays in Eclipse,
they would not be using VJET anymore.

I am not entirely sure why VJET was ever relevant for AJDT in the first
place, so it is safe to remove this line.

Signed-off-by: Alexander Kriegisch <Alexander@Kriegisch.name>